# gym-app-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RankD API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/gyms": {
      "post": {
        "tags": [
          "gyms"
        ],
        "summary": "Create Gym",
        "operationId": "create_gym_gyms_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GymCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymCreatedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/by-code/{code}": {
      "get": {
        "tags": [
          "gyms"
        ],
        "summary": "Get Gym By Code",
        "description": "Preview the gym a join code belongs to (for the \"You're joining X\" screen\nbefore/without signing in) - deliberately public, unlike every other gym\nendpoint. A QR scan happens before login, so this can't require a token;\nit only leaks non-sensitive marketing info (name/city/color, not the gym's\nid), same spirit as the public leaderboard page in the plan's Section 9.\n8 chars of A-Z0-9 is ~2.8e12 codes, so brute-forcing one is impractical.",
        "operationId": "get_gym_by_code_gyms_by_code__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymPreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/{gym_id}": {
      "get": {
        "tags": [
          "gyms"
        ],
        "summary": "Get Gym",
        "operationId": "get_gym_gyms__gym_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/{gym_id}/join-code": {
      "get": {
        "tags": [
          "gyms"
        ],
        "summary": "Get Join Code",
        "description": "The gym's join code, for printing its QR poster. Gym admins and platform\nadmins only - GymOut deliberately omits it so members can't scrape codes.",
        "operationId": "get_join_code_gyms__gym_id__join_code_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinCodeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/{gym_id}/leaderboard": {
      "get": {
        "tags": [
          "gyms"
        ],
        "summary": "Get Leaderboard",
        "operationId": "get_leaderboard_gyms__gym_id__leaderboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GymLeaderboardEntry"
                  },
                  "title": "Response Get Leaderboard Gyms  Gym Id  Leaderboard Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/{gym_id}/exercises/{exercise_id}/weekly-leaderboard": {
      "get": {
        "tags": [
          "gyms"
        ],
        "summary": "Get Weekly Leaderboard",
        "operationId": "get_weekly_leaderboard_gyms__gym_id__exercises__exercise_id__weekly_leaderboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          },
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeeklyLeaderboardEntry"
                  },
                  "title": "Response Get Weekly Leaderboard Gyms  Gym Id  Exercises  Exercise Id  Weekly Leaderboard Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GymCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "brand_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Color"
          }
        },
        "type": "object",
        "required": [
          "name",
          "city"
        ],
        "title": "GymCreate"
      },
      "GymCreatedOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "brand_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Color"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "join_code": {
            "type": "string",
            "title": "Join Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "city",
          "brand_color",
          "created_at",
          "join_code"
        ],
        "title": "GymCreatedOut"
      },
      "GymLeaderboardEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "percentile": {
            "type": "number",
            "title": "Percentile"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "bar_total": {
            "type": "number",
            "title": "Bar Total"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "percentile",
          "user_id",
          "username",
          "bar_total"
        ],
        "title": "GymLeaderboardEntry"
      },
      "GymOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "brand_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Color"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "city",
          "brand_color",
          "created_at"
        ],
        "title": "GymOut",
        "description": "Public shape - deliberately excludes join_code so it can't be looked up\nby anyone who merely knows/guesses a gym's id; it's only handed back once,\nat creation, via GymCreatedOut."
      },
      "GymPreviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "brand_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "city",
          "brand_color"
        ],
        "title": "GymPreviewOut",
        "description": "What someone holding a join code may learn about a gym before joining it,\nsigned in or not: enough to show \"You're joining X\" and, once signed in, to\ncheck existing membership and redirect after joining. The gym's id is a\nUUID primary key, not a capability - every gym-scoped read/write still\ngoes through require_membership regardless of who knows it - so unlike\njoin_code (the actual secret), there's no reason to withhold it here."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JoinCodeOut": {
        "properties": {
          "join_code": {
            "type": "string",
            "title": "Join Code"
          }
        },
        "type": "object",
        "required": [
          "join_code"
        ],
        "title": "JoinCodeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyLeaderboardEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "estimated_1rm_kg": {
            "type": "number",
            "title": "Estimated 1Rm Kg"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "user_id",
          "username",
          "weight_kg",
          "estimated_1rm_kg"
        ],
        "title": "WeeklyLeaderboardEntry"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}